backend/services/restaurants/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
alembic>=1.12.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
